Implement species and mixture as separate submodules
