+++ minplascalc/__init__.py
@@ -1,4 +1,5 @@
 __version__ = '0.7'
 
-from .core import *
+from . import species
+from . import mixture
 from . import readers
