+++ test/unit/test_datafiles.py
@@ -3,14 +3,14 @@ import minplascalc as mpc
 
 
 def test_build_monatomic_species_json():
-    result = mpc.species_from_name('O+')
+    result = mpc.species.from_name('O+')
 
     assert len(result.energylevels) == 275
     assert result.sources is not None
 
 
 def test_build_diatomic_species_json():
-    result = mpc.species_from_name('CO')
+    result = mpc.species.from_name('CO')
     
     assert result.chargenumber == 0
     assert result.sources is not None
